se realizan correciones para poder persistir, crear y loguear un usuario

+++ src/main/java/ar/edu/unnoba/POO/model/QR/controller/AdminController.java
@@ -3,9 +3,11 @@ package ar.edu.unnoba.POO.model.QR.controller;
 import ar.edu.unnoba.POO.model.QR.model.Adminitrador;
 import ar.edu.unnoba.POO.model.QR.service.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
 
@@ -13,15 +15,10 @@ import java.util.List;
 @RequestMapping("/admin")
 public class AdminController {
     private IUserService userService;
+
     @Autowired
     public AdminController(IUserService userService){
-        this.userService= userService;
-    }
-    @GetMapping
-    public String index(Model model){
-        List<Adminitrador> users = userService.getAll();
-        model.addAttribute("admin",users);
-        return "admin/index";
+        this.userService = userService;
     }
 
     @GetMapping("/new")
@@ -30,18 +27,33 @@ public class AdminController {
         return "admin/new";
     }
 
+    @GetMapping
+    public String index(Model model, Authentication authentication){
+        Adminitrador sessionUser = (Adminitrador) authentication.getPrincipal();
+        List<Adminitrador> admin = userService.getAll();
+        model.addAttribute("admin",admin);
+        model.addAttribute("sessionUser",sessionUser);
+        return "admin/index";
+    }
 
     @PostMapping
     public String create(@ModelAttribute Adminitrador user){
         userService.create(user);
-        return "redirect:/admin/new";
+        return "redirect:/admin";
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable("id") Long id){
+    public String delete(@PathVariable("id") Long id,
+                         Authentication authentication,
+                         RedirectAttributes redirectAttributes){
+        Adminitrador sessionUser = (Adminitrador) authentication.getPrincipal();
+        if(sessionUser.getId().equals(id)){
+            redirectAttributes.addFlashAttribute("message","No te podes borrar");
+            return "redirect:/admin";
+        }
         userService.delete(id);
-        return "redirect:/admin/";
+        return "redirect:/admin";
     }
-// metodo  public Admin (Authentication authentication)
+
 
 }
